+++ apps/web/app/(main)/profile/page.tsx
@@ -1,8 +1,21 @@
 import { redirect } from 'next/navigation';
 
-import { getProfile } from '../../../features/profile/api/getProfile';
-import { ProfilePage } from '../../../pages/profile-page';
-import { getCurrentUser } from '../../session';
+import { Navigation, NavigationItem, UserProfileCard } from '@/widgets/profile';
+
+import { getProfile } from '@/features/profile/api/getProfile';
+
+import { getCurrentUser } from '@/app/session';
+
+const pointItems: NavigationItem[] = [
+  { label: '충전내역', href: '/1' },
+  { label: '사용내역', href: '/2' },
+];
+
+const auctionItems: NavigationItem[] = [
+  { label: '판매 중인 물품', href: '/3' },
+  { label: '참여 중인 경매', href: '/4' },
+  { label: '낙찰 받은 물품', href: '/5' },
+];
 
 const Page = async () => {
   const userData = await getCurrentUser();
@@ -13,7 +26,26 @@ const Page = async () => {
 
   const userProfile = await getProfile(userData?.id!);
 
-  return <ProfilePage user={userProfile} />;
+  return (
+    <main className="text-neutral-70" role="main">
+      <h1 className="mb-3 text-base font-semibold">프로필</h1>
+      <UserProfileCard user={userProfile} />
+
+      <section className="mb-7" aria-labelledby="point-management-title">
+        <h2 id="point-management-title" className="mb-3 text-base font-semibold">
+          포인트 내역 관리
+        </h2>
+        <Navigation title="포인트 내역 메뉴" items={pointItems} />
+      </section>
+
+      <section className="mb-7" aria-labelledby="auction-status-title">
+        <h2 id="auction-status-title" className="mb-3 text-base font-semibold">
+          내 경매 현황
+        </h2>
+        <Navigation title="경매 현황 메뉴" items={auctionItems} />
+      </section>
+    </main>
+  );
 };
 
 export default Page;
