+++ apps/web/app/api/chat/list/fetchChatRooms.ts
@@ -1,4 +1,4 @@
-import { supabase } from '../../../lib/supabase/supabase';
+import { supabase } from '../../../../lib/supabase/supabase';
 import { ChatRoom } from '../types';
 
 export const fetchChatRooms = async (userId: string): Promise<ChatRoom[]> => {
