+++ apps/web/app/api/chat/types.ts
