--- apps/web/entities/auction/api/README.md
@@ -1,1 +1,0 @@
-api 폴더입니다
