--- apps/web/entities/auction/model/README.md
@@ -1,1 +1,0 @@
-model 폴더입니다
