+++ apps/web/features/auction/ui/ProductDescriptionInput.tsx
