--- apps/web/features/authentication/ui/README.md
@@ -1,1 +1,0 @@
-ui 폴더입니다
