--- apps/web/features/chat-list/api/README.md
@@ -1,1 +1,0 @@
-api
