+++ apps/web/features/chat-room/model/useMessageSubscription.ts
@@ -2,7 +2,7 @@
 
 import { useEffect, useRef } from 'react';
 
-import { Message } from '../../../entities/chat/types';
+import { Message } from '../../../app/api/chat/types';
 import { supabase } from '../../../lib/supabase/supabase';
 
 export const useMessageSubscription = ({
