+++ apps/web/features/chat-room/ui/ChatMessages.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 
 import { Avatar } from '../../../../../packages/ui/src/design-system/base-components/Avatar/index';
-import { Message, User } from '../../../entities/chat/types';
+import { Message, User } from '../../../app/api/chat/types';
 import { supabase } from '../../../lib/supabase/supabase';
 import { useMessages } from '../model/useFetchMessages';
 import { useMessageSubscription } from '../model/useMessageSubscription';
