+++ apps/web/features/profile/api/update.ts
@@ -1,4 +1,4 @@
-import { supabase } from '../../../lib/supabase/supabase';
+import { supabase } from '@/lib/supabase/supabase';
 
 export const updateProfile = async ({
   id,
