--- apps/web/pages/auction-add-page/index.ts
@@ -1,1 +1,0 @@
-export { AuctionAddPage } from './ui';
