--- apps/web/pages/auction-detail-page/index.ts
@@ -1,1 +1,0 @@
-export { AuctionDetailPage } from './ui';
