--- apps/web/pages/auction-list-page/index.ts
@@ -1,1 +1,0 @@
-export { AuctionListPage } from './ui';
