--- apps/web/pages/charge-history-page/index.ts
@@ -1,1 +1,0 @@
-export { ChargeHistoryPage } from './ui';
