--- apps/web/pages/chatpage.test.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const chatpage.test = () => {
-  return (
-    <div>chatpage.test</div>
-  )
-}
-
-export default chatpage.test
