--- apps/web/pages/main-page/index.ts
@@ -1,1 +1,0 @@
-export { MainPage } from './ui';
