--- apps/web/pages/my-listings-page/index.ts
@@ -1,1 +1,0 @@
-export { MyListingsPage } from './ui';
