--- apps/web/pages/my-listings-page/ui/index.tsx
@@ -1,12 +1,0 @@
-import { mockListings } from '../../../mock/auction';
-import { AuctionListings } from '../../../widgets/auction-listings/ui/AuctionListings';
-
-export const MyListingsPage = () => {
-  return (
-    <main className="text-neutral-70" role="main">
-      <h1 className="mb-3 text-base font-semibold">판매중인 물품</h1>
-
-      <AuctionListings listData={mockListings} />
-    </main>
-  );
-};
