--- apps/web/pages/profile-page/index.ts
@@ -1,1 +1,0 @@
-export { ProfilePage } from './ui';
