--- apps/web/pages/profile-update-page/index.ts
@@ -1,1 +1,0 @@
-export { ProfileForm } from './ui';
