+++ apps/web/tsconfig.json
@@ -5,7 +5,9 @@
       {
         "name": "next"
       }
-    ]
+    ],
+    "baseUrl": ".",
+    "paths": { "@/*": ["*"] }
   },
   "include": [
     "**/*.ts",
